refactored yarn tests to its own py file ; updated rapids.sh to separate services into their own units
