+++ rapids/verify_rapids_dask.py
@@ -2,10 +2,8 @@ import cudf
 import dask_cudf
 import xgboost
 
-from dask.distributed import Client
 import dask.array as da
 import numpy as np
-
 
 def test_rapids():
   # confirm RAPIDS and xgboost are available
@@ -19,21 +17,4 @@ def test_rapids():
   ds = dask_cudf.from_cudf(df['c'], npartitions=2)
   ds.compute()
 
-
-def test_dask_yarn():
-  try:
-    from dask_yarn import YarnCluster
-  except:
-    return
-
-  # Validate dask_yarn configuration
-  cluster = YarnCluster()
-  client = Client(cluster)
-
-  cluster.scale(4)
-  x = da.sum(np.ones(5))
-  x.compute()
-
-
 test_rapids()
-test_dask_yarn()
